+++ CHANGELOG.md
@@ -5,6 +5,7 @@ unreleased
 ----------
 
 - Use version 3.20.0 of braintree-web
+- Update version 4.0.91 of checkout.js
 
 1.4.0
 ------