+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+unreleased
+----------
+
+- Use version 3.20.0 of braintree-web
+- Update browser-detection to v1.5.0
+
 1.4.0
 ------
 - Add `paymentOptionSelected` event